runtime: starting work towards discord classes, implmented the message one
