+++ components/runtime/src/ts/discord/index.ts
@@ -1,0 +1,4 @@
+export * from './dapi';
+export * from './interaction';
+export * from './message';
+export * from '../generated/discord/index';
