+++ components/runtime/src/ts/docs_index.ts
@@ -8,5 +8,5 @@ export * from './script';
 export * from './storage';
 export * from './httpclient';
 export * from './scheduled_tasks';
-export * as Discord from './discord';
+export * as Discord from './discord/index';
 export * as Unstable from './unstable/index';
