--- components/runtime/src/ts/generated.ts
@@ -1,2 +1,0 @@
-export * as DiscordModels from './generated/discord/index';
-export * as Internal from './generated/internal/index';
